colocar logo em todas as telas e adicionar navigation dos fragments ao projeto
