+++ ProjetoToDoList/app/build.gradle.kts
@@ -37,6 +37,10 @@ android {
 
 dependencies {
 
+    //noinspection UseTomlInstead
+    implementation("androidx.navigation:navigation-fragment-ktx:2.8.3")
+    implementation("androidx.navigation:navigation-ui-ktx:2.8.3")
+
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.appcompat)
     implementation(libs.material)
