+++ ProjetoToDoList/app/src/main/java/com/gabreucast/projetotodolist/Fragmets/fragment_home_to_do_list.kt
@@ -1,0 +1,21 @@
+package com.gabreucast.projetotodolist.Fragmets
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.gabreucast.projetotodolist.R
+
+class Hometodolist : Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_home_to_do_list, container, false)
+    }
+
+
+}
