+++ ProjetoToDoList/app/src/main/java/com/gabreucast/projetotodolist/Fragmets/fragment_principal_edit.kt
@@ -1,0 +1,20 @@
+package com.gabreucast.projetotodolist.Fragmets
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.gabreucast.projetotodolist.R
+
+class fragment_principal_edit : Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_principal_edit, container, false)
+    }
+
+}
