+++ ProjetoToDoList/app/src/main/res/layout/fragment_home_to_do_list.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:padding="15dp">
+    android:padding="16dp">
 
     <!-- Fundo do Titulo e Imagem principal -->
     <View
