+++ ProjetoToDoList/app/src/main/res/layout/fragment_principal_edit.xml
@@ -4,16 +4,45 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:padding="12dp"
+    android:padding="16dp"
     tools:context=".UpdateNoteFragment">
+
+    <!-- Fundo do Titulo e Imagem principal -->
+    <View
+        android:id="@+id/blueBackground"
+        style="@style/ViewBackground"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <!-- Título da aplicação -->
+    <androidx.appcompat.widget.AppCompatTextView
+        android:id="@+id/titleTextView"
+        style="@style/TitleText"
+        android:paddingBottom="30dp"
+        android:text="@string/title"
+        android:layout_marginTop="5dp"
+        android:layout_marginStart="5dp"
+        app:layout_constraintStart_toEndOf="@+id/iconImageView"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <!-- Image do App -->
+    <androidx.appcompat.widget.AppCompatImageView
+        android:id="@+id/iconImageView"
+        style="@style/ImageTitulo"
+        android:src="@drawable/baseline_list_alt_24"
+        android:layout_marginStart="10dp"
+        app:layout_constraintStart_toStartOf="@+id/blueBackground"
+        app:layout_constraintTop_toTopOf="parent"/>
 
     <!-- Título da tarefa -->
     <androidx.appcompat.widget.AppCompatTextView
         android:id="@+id/addtaskEditText"
         style="@style/TaskTitleText"
         android:text="@string/editTask"
+        android:layout_marginTop="50dp"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+        app:layout_constraintTop_toBottomOf="@+id/blueBackground" />
 
     <!-- Campo de edição da tarefa -->
     <androidx.appcompat.widget.AppCompatEditText
