+++ ProjetoToDoList/app/src/main/res/values/strings.xml
@@ -26,4 +26,6 @@
 
     <!-- Exemplo de tarefa -->
     <string name="exemploTarefa">Estudar para examen da 5ta feira</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
